+++ uninstall.php
@@ -31,6 +31,7 @@ function seo_campaign_hub_uninstall() {
         'seo_campaign_hub_analytics_ignore_bots',
         'seo_campaign_hub_analytics_ignore_users',
         'seo_campaign_hub_url_replacements',
+        'seo_campaign_hub_text_replacements',
     ];
 
     foreach ( $options as $option ) {